requirements.txt
azure-core==1.0.0
azure-storage-blob==12.0.0
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
idna==2.8
isodate==0.6.0
msrest==0.6.10
oauthlib==3.1.0
paramiko==2.6.0
py4j==0.10.7
pycparser==2.19
PyNaCl==1.3.0
pysftp==0.2.9
pyspark==2.3.2
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
urllib3==1.25.7
